Adds route timeouts to go from particle page to big text page (new) to breathe
