+++ styles/big_text.module.css
@@ -1,0 +1,152 @@
+.container {
+  padding: 0 0rem;
+}
+
+.main {
+  min-height: 100vh;
+  width: 100%;
+  padding: 0;
+  margin: 0;
+  flex: 1;
+  position: absolute;
+  top: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  font-weight: bolder;
+  text-align: center;
+}
+
+.bigText_THIS {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 41vw;
+  z-index: 1;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 1s, 3s;
+  animation-fill-mode: forwards, forwards;
+}
+
+.bigText_IS {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 10vw;
+  z-index: 2;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 3s, 3.2s;
+  animation-fill-mode: forwards, forwards;
+}
+
+.bigText_THE {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 10vw;
+  z-index: 3;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 3.2s, 3.4s;
+  animation-fill-mode: forwards, forwards;
+}
+
+.bigText_2022 {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 35vw;
+  z-index: 4;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 3.4s, 4.2s;
+  animation-fill-mode: forwards, forwards;
+}
+
+.bigText_PERSONAL {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 15vw;
+  z-index: 5;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 4.2s, 5s;
+  animation-fill-mode: forwards, forwards;
+}
+
+/* .bigText_WEBSITE {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 20vw;
+  z-index: 6;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 1s, 3s;
+  animation-fill-mode: forwards, forwards;
+} */
+
+.bigText_OF {
+  opacity: 0;
+  position: fixed;
+  font-family: Inter;
+  font-size: 10vw;
+  z-index: 7;
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 5s, 5.2s;
+  animation-fill-mode: forwards, forwards;
+}
+
+.bigText_CHLOE_NOTT {
+  position: fixed;
+  z-index: 8;
+}
+
+.bigText_CHLOE {
+  opacity: 0;
+  position: relative;
+  font-family: Inter;
+  font-size: 26vw;
+  line-height: 90%;
+  /* z-index: 8; */
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 5.2s, 7.2s;
+  animation-fill-mode: forwards, forwards;
+}
+
+.bigText_NOTT {
+  opacity: 0;
+  position: relative;
+  font-family: Inter;
+  font-size: 31vw;
+  line-height: 90%;
+  /* z-index: 9; */
+  animation: fade_in, fade_out;
+  animation-duration: 0s, 0s;
+  animation-delay: 6.2s, 7.2s;
+  animation-fill-mode: forwards, forwards;
+}
+
+@keyframes fade_in {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+@keyframes fade_out {
+  0% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 0;
+  }
+}
